backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==3.6.4.0
celery==5.2.2
certifi==2023.7.22
chardet==3.0.4
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-cors-headers==4.2.0
django-countries==7.5.1
django-extensions==3.2.3
django-import-export==3.2.0
django-jazzmin==2.6.0
django-language-field==0.0.3
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==2.8
kombu==5.3.1
Languages==1.0.0
MarkupPy==1.14
mtnmomo==3.0.1
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.18
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
reportlab==4.0.4
requests==2.21.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.24.3
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
